backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
b2sdk
pydantic
requests
python-multipart
yt-dlp
